% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exTREEmaTIME_Main.R
\name{exTREEmaTIMEcalc}
\alias{exTREEmaTIMEcalc}
\title{performs core analysis and writes max age tree and min age tree to file
writes branch specific rates in max age tree and min age tree to file}
\usage{
exTREEmaTIMEcalc(
  tree,
  rmax,
  rmin,
  n_max_constraints,
  max_constraint_clade,
  max_constraint_age,
  n_min_constraints,
  min_constraint_clade,
  min_constraint_age,
  calibration_implementation_precision,
  sample_time
)
}
\description{
performs core analysis and writes max age tree and min age tree to file
writes branch specific rates in max age tree and min age tree to file
}
