% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exTREEmaTIME_Core.r
\name{exTREEmaTIMEcore}
\alias{exTREEmaTIMEcore}
\title{performs core analysis and writes max age tree and min age tree to file
writes branch specific rates in max age tree and min age tree to file}
\usage{
exTREEmaTIMEcore(
  tree,
  rmax,
  rmin,
  noise_level,
  n_max_constraints,
  max_constraints_clade,
  max_constraints_ages,
  n_min_constraints,
  min_constraints_clade,
  min_constraints_ages,
  calibration_implementation_precision,
  tip,
  sample_time
)
}
\description{
performs core analysis and writes max age tree and min age tree to file
writes branch specific rates in max age tree and min age tree to file
}
