% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exTREEmaTIME_Main.r
\name{exTREEmaTIMEmain}
\alias{exTREEmaTIMEmain}
\title{main input function for used
enables parameterisation of rmin and rmax (including SetAutoRates) and enables full exTREEmaTIME analysis}
\usage{
exTREEmaTIMEmain(
  tree,
  auto_rates,
  auto_rates_type,
  max_rate_unit,
  min_rate_unit,
  rmax,
  rmin,
  noise_level,
  root_max,
  root_min,
  n_max_constraints,
  max_constraints_clade,
  max_constraints_ages,
  n_min_constraints,
  min_constraints_clade,
  min_constraints_ages,
  calibration_implementation_precision,
  tip,
  sample_time,
  output_directory
)
}
\description{
main input function for used
enables parameterisation of rmin and rmax (including SetAutoRates) and enables full exTREEmaTIME analysis
}
