% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exTREEmaTIME_SetAutoRates.R
\name{SetAutoRates}
\alias{SetAutoRates}
\title{calculates plausible values for rmax and rmin for subsequent use in UncertainTree
increase the unit until UncertainTreeCalc reaches completion}
\usage{
SetAutoRates(tree, root_min, root_max, unit)
}
\description{
calculates plausible values for rmax and rmin for subsequent use in UncertainTree
increase the unit until UncertainTreeCalc reaches completion
}
